+++ examples/anu_lms_examples/src/EventSubscriber/AnuLmsExamplesSubscriber.php
@@ -5,6 +5,7 @@ namespace Drupal\anu_lms_examples\EventSubscriber;
 use Drupal\Core\Messenger\MessengerInterface;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 use Drupal\anu_lms\Event\CoursesPageDataGeneratedEvent;
+use Drupal\anu_lms\Event\LessonPageDataGeneratedEvent;
 
 /**
  * Anu LMS Examples event subscriber.
@@ -34,7 +35,7 @@ class AnuLmsExamplesSubscriber implements EventSubscriberInterface {
    * @param \Drupal\anu_lms\Event\CoursesPageDataGeneratedEvent $event
    *   Response event.
    */
-  public function onDataGenerated(CoursesPageDataGeneratedEvent $event) {
+  public function onCoursesDataGenerated(CoursesPageDataGeneratedEvent $event) {
     $data = $event->getPageData();
     $node = $event->getNode();
     $data['additional-example-data'] = 'test';
@@ -43,11 +44,26 @@ class AnuLmsExamplesSubscriber implements EventSubscriberInterface {
   }
 
   /**
+   * Event handler.
+   *
+   * @param \Drupal\anu_lms\Event\LessonPageDataGeneratedEvent $event
+   *   Response event.
+   */
+  public function onLessonDataGenerated(LessonPageDataGeneratedEvent $event) {
+    $data = $event->getPageData();
+    $node = $event->getNode();
+    $data['lesson-example-data'] = 'test';
+    $this->messenger->addStatus('Data was generated for lesson "' . $node->label() . '"');
+    $event->setPageData($data);
+  }
+
+  /**
    * {@inheritdoc}
    */
   public static function getSubscribedEvents() {
     return [
-      'anu_lms_courses_page_data_generated' => ['onDataGenerated'],
+      'anu_lms_courses_page_data_generated' => ['onCoursesDataGenerated'],
+      'anu_lms_lesson_page_data_generated' => ['onLessonDataGenerated'],
     ];
   }