+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Refactored quiz submission API endpoint for better code readability, improved validation & access checks.
  - Refactored TODO list. Removed completed item.
  - Updated "ansi-regex" package to fix security vulnerabilities.
+ - Added an event to be able to modify output of lesson page.
 
 ## [2.4.0]
  - Quizzes submodule is optional and the core module has no longer a dependency on it.